Fix overflow when calculating history values I thought this would be more impactful
TC: 2+0.1 Elo: 2.32 +/- 24.06 Games: 600
TC: 30+0.5 Elo: 8.69 +/- 27.28 Games: 400

+++ src/search.rs
@@ -682,7 +682,7 @@ impl Board {
         let history_color_value = if self.white_to_move { 0 } else { 1 };
 
         let current_value = &mut history_table[history_color_value][piece_type - 1][m.to() as usize];
-        let clamped_bonus = bonus.clamp(-MOVE_SCORE_HISTORY_MAX, MOVE_SCORE_HISTORY_MAX);
-        *current_value += clamped_bonus - ((*current_value) * clamped_bonus.abs() / MOVE_SCORE_HISTORY_MAX);
+        let clamped_bonus = (bonus as i32).clamp(-MOVE_SCORE_HISTORY_MAX, MOVE_SCORE_HISTORY_MAX);
+        *current_value += (clamped_bonus - ((*current_value as i32) * clamped_bonus.abs() / MOVE_SCORE_HISTORY_MAX)) as i16;
     }
 }
